[SPOT-281][REFACTOR] 회원 정보 관련 기능 유즈 케이스로 분리
