+++ src/main/java/com/example/spot/member/application/refactor/impl/MemberInfoServiceImpl.java
@@ -1,0 +1,49 @@
+package com.example.spot.member.application.refactor.impl;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.example.spot.common.api.code.status.ErrorStatus;
+import com.example.spot.common.api.exception.handler.MemberHandler;
+import com.example.spot.member.application.refactor.MemberInfoService;
+import com.example.spot.member.domain.Member;
+import com.example.spot.member.domain.MemberRepository;
+import com.example.spot.member.presentation.dto.MemberRequestDTO;
+import com.example.spot.member.presentation.dto.MemberResponseDTO;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class MemberInfoServiceImpl implements MemberInfoService {
+
+	private final MemberRepository memberRepository;
+	/**
+	 * 회원의 프로필을 업데이트 합니다.
+	 * @param memberId 변경할 회원 ID
+	 * @param requestDTO 변경할 회원 정보
+	 * @return 변경 된 회원 ID와 변경 시간
+	 * @throws MemberHandler 회원을 찾을 수 없을 경우
+	 */
+	@Override
+	public MemberResponseDTO.MemberUpdateDTO updateProfile(Long memberId, MemberRequestDTO.MemberUpdateDTO requestDTO) {
+		// 회원 조회
+		Member member = memberRepository.findById(memberId)
+				.orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
+
+		// 회원 정보 업데이트
+		member.updateInfo(requestDTO);
+
+		// 회원 정보 저장
+		memberRepository.save(member);
+
+		// 업데이트된 회원 정보 반환
+		return MemberResponseDTO.MemberUpdateDTO.builder()
+				.memberId(member.getId())
+				.updatedAt(member.getUpdatedAt())
+				.build();
+	}
+}
