+++ src/main/java/com/example/spot/member/presentation/controller/legacy/MemberController.java
@@ -33,6 +33,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import static com.example.spot.auth.presentation.dto.google.GoogleExampleResponse.EXAMPLE_RESPONSE;
 
+@Deprecated
 @RestController
 @RequestMapping("/spot")
 @RequiredArgsConstructor
@@ -68,22 +69,6 @@ public class MemberController {
         return ApiResponse.onSuccess(SuccessStatus._MEMBER_CREATED, dto);
     }
 
-
-
-    @Tag(name = "회원 관리 API", description = "회원 관리 API")
-    @PostMapping("/members/user-info")
-    @Operation(summary = "[회원 정보 업데이트] 개인 정보 입력 및 수정",
-        description = """
-            ## [회원 정보 업데이트] 해당하는 회원의 개인 정보를 입력 및 수정 합니다.
-            업데이트 할 회원의 정보를 입력 받습니다.
-            대상 회원의 식별 아이디와 수정 시각이 반환 됩니다. 
-            """,
-        security = @SecurityRequirement(name = "accessToken"))
-    public ApiResponse<MemberUpdateDTO> updateMemberInfo(
-        @RequestBody @Valid MemberRequestDTO.MemberUpdateDTO requestDTO){
-        MemberUpdateDTO memberUpdateDTO = memberService.updateProfile(SecurityUtils.getCurrentUserId(), requestDTO);
-        return ApiResponse.onSuccess(SuccessStatus._MEMBER_INFO_UPDATE, memberUpdateDTO);
-    }
